refactor: extract preload management into a mixin
The `# Preload management` banner block moves out of PdsFile into _PreloadMixin
in _preload.py: get_permanent_values, load_volume_info,
cache_category_merged_dirs, preload (with its nested _preload_dir) and
cache_lifetime. 419 lines lifted by line span, byte-identical, banner included.
The block ends at cache_lifetime: new_merged_dir, new_index_row_pdsfile, copy
and __repr__ sit under the same banner but are on PR-22's stay-list, so the
boundary is a symbol boundary and not a banner boundary.

_PreloadMixin takes its alphabetical slot between _OpusMixin and _ShelfMixin.
No class-level assignment moves; CACHE, MEMCACHE_PORT, DEFAULT_CACHING,
LOCAL_PRELOADED, PRELOAD_TRIES and DICTIONARY_CACHE_LIMIT stay on PdsFile,
where the moved methods read and write them through cls.

The keep-green edits the move requires:

* _preload.py gains the imports the moved bodies use -- os, time, pdscache and
  pdsviewable, _clean_abspath and _clean_join, and the try/except that sets
  HAS_PYLIBMC, which exists only for preload's memcached branch;
* pdsfile.py binds five names it no longer references itself in the PEP-484
  redundant-alias form, because each is either a frozen member of its public
  surface (time, HAS_PYLIBMC, pause_caching, resume_caching) or reachable as
  pdsfile.pdsfile.<name> (_clean_abspath);
* pyproject.toml gains _preload.py's per-file-ignores entry, a split of
  pdsfile.py's: E501, E701, F841, RUF005, UP015, UP031, every one already
  forgiven for these same lines. UP015's only occurrence in pdsfile.py was in
  load_volume_info, so pdsfile.py's entry drops it.

pylibmc itself is not carried back. It is not a manifest member, and a stub on
PYTHONPATH shows that where it is importable the freeze gate is already red,
because the dumper then records it as an extra name. It resolves as
pdsfile._preload.pylibmc.
