+++ pyproject.toml
@@ -167,6 +167,7 @@ extend-ignore = ["PT011", "SIM105", "SIM108"]
 "src/pdsfile/_local_fs.py" = ["B007", "B905", "E701", "SIM103", "SIM118", "UP024"]
 "src/pdsfile/_opus.py" = ["RUF005", "UP024"]
 "src/pdsfile/_path_utils.py" = ["E701", "F841"]
+"src/pdsfile/_preload.py" = ["E501", "E701", "F841", "RUF005", "UP015", "UP031"]
 "src/pdsfile/_shelves.py" = ["B904", "B905", "F841", "RUF059", "SIM103", "UP024", "UP031"]
 "src/pdsfile/_sorting.py" = ["E701", "RUF005"]
 "src/pdsfile/holdings_maintenance/pds3/crlf.py" = ["PT028"]
@@ -220,7 +221,7 @@ extend-ignore = ["PT011", "SIM105", "SIM108"]
 "src/pdsfile/pds4file/rules/cassini_vims.py" = ["E501", "I001", "N801", "RUF012"]
 "src/pdsfile/pds4file/rules/uranus_occs_earthbased.py" = ["B007", "E501", "I001", "N801", "RUF012"]
 "src/pdsfile/pdscache.py" = ["A001", "E701", "E721", "F401", "F541", "F841", "I001", "RUF015", "RUF059", "SIM102", "UP004", "UP031"]
-"src/pdsfile/pdsfile.py" = ["B904", "C405", "E501", "E701", "E713", "E721", "F841", "I001", "N806", "RUF005", "RUF012", "SIM102", "SIM114", "SIM118", "UP004", "UP015", "UP024", "UP031"]
+"src/pdsfile/pdsfile.py" = ["B904", "C405", "E501", "E701", "E713", "E721", "F841", "I001", "N806", "RUF005", "RUF012", "SIM102", "SIM114", "SIM118", "UP004", "UP024", "UP031"]
 "src/pdsfile/pdsviewable.py" = ["B006", "B007", "B020", "C405", "E701", "E721", "F401", "I001", "RUF015", "RUF059", "UP004", "UP024"]
 "src/pdsfile/tools/show_opus_products.py" = ["I001"]
 "tests/conftest.py" = ["F401", "I001", "N806"]
